split up WireInfo etc. into its own file.
